ASN IP 주소 파싱/저장 기능을 추가함 OID 의 64bit 정수 파싱 기능을 추가함

+++ SnmpParser/AsnOid.h
@@ -39,10 +39,11 @@ public:
 
 	virtual bool GetString( std::string & strValue );
 
+	/** OID */
 	std::string m_strValue;
 
 private:
-	static void SetOidEntry( char * pszPacket, int iPacketSize, int iValue, int & iPos );
+	static void SetOidEntry( char * pszPacket, int iPacketSize, uint64_t iValue, int & iPos );
 };
 
 #endif
